color: split out Colour parsing code
Move the pure Colour parsing functions into a separate file
so that they can be tested.

- parse_attr_spec()
- parse_color_colornnn()
- parse_color_name()
- parse_color_namedcolor()
- parse_color_pair()
- parse_color_prefix()
- parse_color_rrggbb()
