+++ Makefile.autosetup
@@ -182,7 +182,7 @@ $(PWD)/browser:
 LIBCOLOR=	libcolor.a
 LIBCOLOROBJS=	color/ansi.o color/attr.o color/color.o color/command.o \
 		color/curses.o color/merged.o color/notify.o \
-		color/parse_ansi.o color/quoted.o \
+		color/parse_ansi.o color/parse_color.o color/quoted.o \
 		color/regex.o color/simple.o
 @if USE_DEBUG_COLOR
 LIBCOLOROBJS+=	color/debug.o
